requirements.txt
alembic==1.7.5
anyio==3.5.0
asgiref==3.4.1
click==8.0.3
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.71.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pydantic==1.9.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.17.1
typing-extensions==4.0.1
uuid==1.30
uvicorn==0.17.0
zipp==3.7.0
